+++ src/Type/Definition/BooleanType.php
@@ -21,23 +21,15 @@ class BooleanType extends ScalarType
     public $description = 'The `Boolean` scalar type represents `true` or `false`.';
 
     /**
-     * Coerce the given value to a boolean.
+     * Serialize the given value to a boolean.
      *
      * The GraphQL spec leaves this up to the implementations, so we just do what
      * PHP does natively to make this intuitive for developers.
      *
      * @param mixed $value
-     *
-     * @throws Error
      */
     public function serialize($value) : bool
     {
-        if (is_array($value)) {
-            throw new Error(
-                'Boolean cannot represent an array value: ' . Utils::printSafe($value)
-            );
-        }
-
         return (bool) $value;
     }
 
@@ -54,7 +46,7 @@ class BooleanType extends ScalarType
             return $value;
         }
 
-        throw new Error('Cannot represent value as boolean: ' . Utils::printSafe($value));
+        throw new Error('Boolean cannot represent a non boolean value: ' . Utils::printSafe($value));
     }
 
     /**