+++ src/Type/Definition/IDType.php
@@ -39,22 +39,12 @@ When expected as an input type, any string (such as `"4"`) or integer
      */
     public function serialize($value)
     {
-        if ($value === true) {
-            return 'true';
-        }
-        if ($value === false) {
-            return 'false';
-        }
-        if ($value === null) {
-            return 'null';
-        }
-        if (is_array($value)) {
-            throw new Error(
-                'ID cannot represent an array value: ' . Utils::printSafe($value)
-            );
-        }
-        if (! is_scalar($value) && (! is_object($value) || ! method_exists($value, '__toString'))) {
-            throw new Error('ID cannot represent non scalar value: ' . Utils::printSafe($value));
+        $canCast = is_string($value)
+            || is_int($value)
+            || (is_object($value) && method_exists($value, '__toString'));
+
+        if (! $canCast) {
+            throw new Error('ID cannot represent value: ' . Utils::printSafe($value));
         }
 
         return (string) $value;
@@ -72,13 +62,7 @@ When expected as an input type, any string (such as `"4"`) or integer
         if (is_string($value) || is_int($value)) {
             return (string) $value;
         }
-        if (is_array($value)) {
-            throw new Error(
-                'ID cannot represent an array value: ' . Utils::printSafe($value)
-            );
-        }
-
-        throw new Error('Cannot represent value as ID: ' . Utils::printSafe($value));
+        throw new Error('ID cannot represent value: ' . Utils::printSafe($value));
     }
 
     /**