+++ src/Type/Definition/IntType.php
@@ -10,9 +10,12 @@ use GraphQL\Language\AST\IntValueNode;
 use GraphQL\Language\AST\Node;
 use GraphQL\Utils\Utils;
 use function floatval;
+use function floor;
 use function intval;
 use function is_array;
 use function is_bool;
+use function is_float;
+use function is_int;
 use function is_numeric;
 use function sprintf;
 
@@ -43,53 +46,28 @@ values. Int can represent values between -(2^31) and 2^31 - 1. ';
      */
     public function serialize($value)
     {
-        return $this->coerceInt($value);
-    }
-
-    /**
-     * @param mixed $value
-     *
-     * @return int
-     */
-    private function coerceInt($value)
-    {
-        if (is_array($value)) {
-            throw new Error(
-                sprintf('Int cannot represent an array value: %s', Utils::printSafe($value))
-            );
+        // Fast path for 90+% of cases:
+        if (is_int($value) && $value <= self::MAX_INT && $value >= self::MIN_INT) {
+            return $value;
         }
 
-        if ($value === '') {
-            throw new Error(
-                'Int cannot represent non-integer value: (empty string)'
-            );
-        }
+        $float = is_numeric($value) || is_bool($value) ? floatval($value) : null;
 
-        if (! is_numeric($value) && ! is_bool($value)) {
+        if ($float === null || floor($float) !== $float) {
             throw new Error(
                 'Int cannot represent non-integer value: ' .
                 Utils::printSafe($value)
             );
         }
 
-        $num = floatval($value);
-        if ($num > self::MAX_INT || $num < self::MIN_INT) {
+        if ($float > self::MAX_INT || $float < self::MIN_INT) {
             throw new Error(
                 'Int cannot represent non 32-bit signed integer value: ' .
                 Utils::printSafe($value)
             );
         }
-        $int = intval($num);
-        // int cast with == used for performance reasons
-        // phpcs:ignore
-        if ($int != $num) {
-            throw new Error(
-                'Int cannot represent non-integer value: ' .
-                Utils::printSafe($value)
-            );
-        }
 
-        return $int;
+        return intval($float);
     }
 
     /**
@@ -101,7 +79,23 @@ values. Int can represent values between -(2^31) and 2^31 - 1. ';
      */
     public function parseValue($value)
     {
-        return $this->coerceInt($value);
+        $isInt = is_int($value) || (is_float($value) && floor($value) === $value);
+
+        if (! $isInt) {
+            throw new Error(
+                'Int cannot represent non-integer value: ' .
+                Utils::printSafe($value)
+            );
+        }
+
+        if ($value > self::MAX_INT || $value < self::MIN_INT) {
+            throw new Error(
+                'Int cannot represent non 32-bit signed integer value: ' .
+                Utils::printSafe($value)
+            );
+        }
+
+        return intval($value);
     }
 
     /**