+++ tests/Type/ScalarSerializationTest.php
@@ -5,14 +5,13 @@ declare(strict_types=1);
 namespace GraphQL\Tests\Type;
 
 use GraphQL\Error\Error;
-use GraphQL\Type\Definition\IDType;
-use GraphQL\Type\Definition\StringType;
+use GraphQL\Tests\Type\TestClasses\CanCastToString;
+use GraphQL\Tests\Type\TestClasses\ObjectIdStub;
 use GraphQL\Type\Definition\Type;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 use function acos;
 use function log;
-use function sprintf;
 
 class ScalarSerializationTest extends TestCase
 {
@@ -34,114 +33,38 @@ class ScalarSerializationTest extends TestCase
         self::assertSame(1, $intType->serialize(true));
     }
 
-    public function testSerializesOutputIntCannotRepresentFloat1() : void
+    public function badIntValues()
+    {
+        return [
+            [0.1, 'Int cannot represent non-integer value: 0.1'],
+            [1.1, 'Int cannot represent non-integer value: 1.1'],
+            [-1.1, 'Int cannot represent non-integer value: -1.1'],
+            ['-1.1', 'Int cannot represent non-integer value: -1.1'],
+            [9876504321, 'Int cannot represent non 32-bit signed integer value: 9876504321'],
+            [-9876504321, 'Int cannot represent non 32-bit signed integer value: -9876504321'],
+            [1e100, 'Int cannot represent non 32-bit signed integer value: 1.0E+100'],
+            [-1e100, 'Int cannot represent non 32-bit signed integer value: -1.0E+100'],
+            [log(0), 'Int cannot represent non 32-bit signed integer value: -INF'],
+            [acos(8), 'Int cannot represent non-integer value: NAN'],
+            ['one', 'Int cannot represent non-integer value: one'],
+            ['', 'Int cannot represent non-integer value: (empty string)'],
+            [[5], 'Int cannot represent non-integer value: [5]'],
+        ];
+    }
+
+    /**
+     * @throws Error
+     *
+     * @dataProvider badIntValues
+     */
+    public function testSerializesOutputAsIntErrors($value, $expectedError) : void
     {
         // The GraphQL specification does not allow serializing non-integer values
         // as Int to avoid accidental data loss.
         $intType = Type::int();
         $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non-integer value: 0.1');
-        $intType->serialize(0.1);
-    }
-
-    public function testSerializesOutputIntCannotRepresentFloat2() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non-integer value: 1.1');
-        $intType->serialize(1.1);
-    }
-
-    public function testSerializesOutputIntCannotRepresentNegativeFloat() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non-integer value: -1.1');
-        $intType->serialize(-1.1);
-    }
-
-    public function testSerializesOutputIntCannotRepresentNumericString() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non-integer value: -1.1');
-        $intType->serialize('-1.1');
-    }
-
-    public function testSerializesOutputIntCannotRepresentBiggerThan32Bits() : void
-    {
-        // Maybe a safe PHP int, but bigger than 2^32, so not
-        // representable as a GraphQL Int
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non 32-bit signed integer value: 9876504321');
-        $intType->serialize(9876504321);
-    }
-
-    public function testSerializesOutputIntCannotRepresentLowerThan32Bits() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non 32-bit signed integer value: -9876504321');
-        $intType->serialize(-9876504321);
-    }
-
-    public function testSerializesOutputIntCannotRepresentBiggerThanSigned32Bits() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non 32-bit signed integer value: 1.0E+100');
-        $intType->serialize(1e100);
-    }
-
-    public function testSerializesOutputIntCannotRepresentLowerThanSigned32Bits() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non 32-bit signed integer value: -1.0E+100');
-        $intType->serialize(-1e100);
-    }
-
-    public function testSerializesOutputIntCannotRepresentInfinity() : void
-    {
-        $intType  = Type::int();
-        $infinity = log(0);
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non 32-bit signed integer value: -INF');
-        $intType->serialize($infinity);
-    }
-
-    public function testSerializesOutputIntCannotRepresentNaN() : void
-    {
-        $intType = Type::int();
-        $nan     = acos(8);
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non-integer value: NAN');
-        $intType->serialize($nan);
-    }
-
-    public function testSerializesOutputIntCannotRepresentString() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non-integer value: one');
-        $intType->serialize('one');
-    }
-
-    public function testSerializesOutputIntCannotRepresentEmptyString() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent non-integer value: (empty string)');
-        $intType->serialize('');
-    }
-
-    public function testSerializesOutputIntCannotRepresentArray() : void
-    {
-        $intType = Type::int();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Int cannot represent an array value: [5]');
-        $intType->serialize([5]);
+        $this->expectExceptionMessage($expectedError);
+        $intType->serialize($value);
     }
 
     /**
@@ -163,112 +86,64 @@ class ScalarSerializationTest extends TestCase
         self::assertSame(1.0, $floatType->serialize(true));
     }
 
-    public function testSerializesOutputFloatCannotRepresentString() : void
-    {
-        $floatType = Type::float();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Float cannot represent non numeric value: one');
-        $floatType->serialize('one');
-    }
-
-    public function testSerializesOutputFloatCannotRepresentEmptyString() : void
-    {
-        $floatType = Type::float();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Float cannot represent non numeric value: (empty string)');
-        $floatType->serialize('');
-    }
-
-    public function testSerializesOutputFloatCannotRepresentInfinity() : void
-    {
-        $floatType = Type::float();
-        $infinity  = log(0);
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Float cannot represent non numeric value: -INF');
-        $floatType->serialize($infinity);
-    }
-
-    public function testSerializesOutputFloatCannotRepresentNaN() : void
-    {
-        $floatType = Type::float();
-        $nan       = acos(8);
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Float cannot represent non numeric value: NAN');
-        $floatType->serialize($nan);
-    }
-
-    public function testSerializesOutputFloatCannotRepresentArray() : void
-    {
-        $floatType = Type::float();
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage('Float cannot represent an array value: [5]');
-        $floatType->serialize([5]);
-    }
-
-    public function stringLikeTypes()
+    public function badFloatValues()
     {
         return [
-            [ Type::string() ],
-            [ Type::id() ],
+            ['one', 'Float cannot represent non numeric value: one'],
+            ['', 'Float cannot represent non numeric value: (empty string)'],
+            [log(0), 'Float cannot represent non numeric value: -INF'],
+            [acos(8), 'Float cannot represent non numeric value: NAN'],
+            [[5], 'Float cannot represent non numeric value: [5]'],
         ];
     }
 
     /**
-     * @see it('serializes output as String')
+     * @throws Error
      *
-     * @param StringType|IDType $stringType
-     *
-     * @dataProvider stringLikeTypes
+     * @dataProvider badFloatValues
      */
-    public function testSerializesOutputAsString($stringType) : void
+    public function testSerializesOutputFloatErrors($value, $expectedError) : void
     {
+        $floatType = Type::float();
+        $this->expectException(Error::class);
+        $this->expectExceptionMessage($expectedError);
+        $floatType->serialize($value);
+    }
+
+    /**
+     * @see it('serializes output as String')
+     */
+    public function testSerializesOutputAsString() : void
+    {
+        $stringType = Type::string();
         self::assertSame('string', $stringType->serialize('string'));
         self::assertSame('1', $stringType->serialize(1));
         self::assertSame('-1.1', $stringType->serialize(-1.1));
-        self::assertSame('true', $stringType->serialize(true));
-        self::assertSame('false', $stringType->serialize(false));
-        self::assertSame('null', $stringType->serialize(null));
-        self::assertSame('2', $stringType->serialize(new ObjectIdStub(2)));
+        self::assertSame('1', $stringType->serialize(true));
+        self::assertSame('', $stringType->serialize(false));
+        self::assertSame('', $stringType->serialize(null));
+        self::assertSame('foo', $stringType->serialize(new CanCastToString('foo')));
+    }
+
+    public function badStringValues()
+    {
+        return [
+            [[1], 'String cannot represent value: [1]'],
+            [new stdClass(), 'String cannot represent value: instance of stdClass'],
+        ];
     }
 
     /**
-     * @param StringType|IDType $stringType
-     *
      * @throws Error
      *
-     * @dataProvider stringLikeTypes
+     * @dataProvider badStringValues
      */
-    public function testSerializesOutputStringsCannotRepresentArray($stringType) : void
+    public function testSerializesOutputStringErrors($value, $expectedError) : void
     {
+        $stringType = Type::string();
         $this->expectException(Error::class);
-        $this->expectExceptionMessage(sprintf('%s cannot represent an array value: [1]', $stringType->name));
-        $stringType->serialize([1]);
-    }
-
-    /**
-     * @param StringType|IDType $stringType
-     *
-     * @dataProvider stringLikeTypes
-     */
-    public function testSerializesOutputStringsCannotRepresentObject($stringType) : void
-    {
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage(sprintf('%s cannot represent non scalar value: instance of stdClass', $stringType->name));
-        $stringType->serialize(new stdClass());
-    }
-
-    /**
-     * @param StringType|IDType $stringType
-     *
-     * @throws Error
-     *
-     * @dataProvider stringLikeTypes
-     */
-    public function testSerializesOutputStringCannotRepresentArray($stringType) : void
-    {
-        $this->expectException(Error::class);
-        $this->expectExceptionMessage(sprintf('%s cannot represent an array value: [5]', $stringType->name));
-        $stringType->serialize([5]);
+        $this->expectExceptionMessage($expectedError);
+        $stringType->serialize($value);
     }
 
     /**
@@ -289,11 +164,42 @@ class ScalarSerializationTest extends TestCase
         self::assertFalse($boolType->serialize(''));
     }
 
-    public function testSerializesOutputBooleanCannotRepresentArray() : void
+    /**
+     * @see it('serializes output as ID')
+     */
+    public function testSerializesOutputAsID() : void
     {
-        $boolType = Type::boolean();
+        $idType = Type::id();
+
+        self::assertSame('string', $idType->serialize('string'));
+        self::assertSame('false', $idType->serialize('false'));
+        self::assertSame('', $idType->serialize(''));
+        self::assertSame('1', $idType->serialize('1'));
+        self::assertSame('0', $idType->serialize('0'));
+        self::assertSame('1', $idType->serialize(1));
+        self::assertSame('0', $idType->serialize(0));
+        self::assertSame('2', $idType->serialize(new ObjectIdStub(2)));
+    }
+
+    public function badIDValues()
+    {
+        return [
+            [new stdClass(), 'ID cannot represent value: instance of stdClass'],
+            [true, 'ID cannot represent value: true'],
+            [false, 'ID cannot represent value: false'],
+            [-1.1, 'ID cannot represent value: -1.1'],
+            [['abc'], 'ID cannot represent value: ["abc"]'],
+        ];
+    }
+
+    /**
+     * @dataProvider badIDValues
+     */
+    public function testSerializesOutputAsIDError($value, $expectedError)
+    {
+        $idType = Type::id();
         $this->expectException(Error::class);
-        $this->expectExceptionMessage('Boolean cannot represent an array value: [5]');
-        $boolType->serialize([5]);
+        $this->expectExceptionMessage($expectedError);
+        $idType->serialize($value);
     }
 }